Initial HSQLUtil implementation.

+++ code/br.cin.ufpe.reviewer.persistence/src/br/cin/ufpe/reviewer/persistence/dao/literaturereview/ILiteratureReviewDAO.java
@@ -3,6 +3,6 @@ package br.cin.ufpe.reviewer.persistence.dao.literaturereview;
 import br.cin.ufpe.reviewer.model.literaturereview.LiteratureReview;
 import br.cin.ufpe.reviewer.persistence.dao.IDAO;
 
-public interface ILiteratureReviewDAO extends IDAO<LiteratureReview, Long> {
+public interface ILiteratureReviewDAO extends IDAO<LiteratureReview, Integer> {
 
 }
