+++ code/br.cin.ufpe.reviewer.persistence/src/br/cin/ufpe/reviewer/persistence/dao/literaturereview/LiteratureReviewJPADAO.java
@@ -5,7 +5,7 @@ import javax.persistence.EntityManager;
 import br.cin.ufpe.reviewer.model.literaturereview.LiteratureReview;
 import br.cin.ufpe.reviewer.persistence.dao.JPADAO;
 
-public class LiteratureReviewJPADAO extends JPADAO<LiteratureReview, Long> implements ILiteratureReviewDAO {
+public class LiteratureReviewJPADAO extends JPADAO<LiteratureReview, Integer> implements ILiteratureReviewDAO {
 
 	public LiteratureReviewJPADAO() {
 		super(LiteratureReview.class);
